Add new tool : lists content of a path
class listLocals provides capability of listing contents of a directory.
Takes single argument `path`. It is a path to the directory whose contents is going
to be returned on call to lsdir() method.

method lsdir() returns a items dictionary of a directory. Takes single optional
`path` to change the default path which was set on initialization by user.